% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catch.R
\name{get_catch}
\alias{get_catch}
\title{Get species catches per tow from sGSL September surveys}
\usage{
get_catch(
  species,
  years,
  adj.ratio = TRUE,
  adj.distance = TRUE,
  adj.vessel = TRUE,
  adj.diel = TRUE,
  experiment.code = c(1, 5),
  cutoff.length = NULL,
  adj.repeat.sets = TRUE
)
}
\arguments{
\item{species}{Which RV species code. Currently limited to certain species if applying vessel and diel corrections (10,12,23,40,41,42,43,60,70).}

\item{years}{Which year(s) to return. Current minimum to apply all corrections is 1993.}

\item{adj.ratio}{Adjust for sampling ratio? Default is TRUE.}

\item{adj.distance}{Adjust for target tow distance? Default is TRUE.}

\item{adj.vessel}{Adjust for vessel effects? Default is TRUE.}

\item{adj.diel}{Adjust for diel effects? Default is TRUE. Note that diel adjustments are only made after vessel adjustments to Teleost.}

\item{experiment.code}{Restrict tow experiment type. See gulf::experiment.str(). Default is c(1,5) ("stratified random survey set", "comparative fishing experiment").}

\item{cutoff.length}{Include estimates below and above a cutoff length. Output columns labeled "small" are estimates below the cutoff length. Output columns labeled "big" are estimates at and above the cutoff length. Default is NULL.}

\item{adj.repeat.sets}{Adjust for repeat sets? Identifies repeat sets (e.g., during comparative tows), and retains only the first set. First sets are retained in order to preserve tow information (e.g., spatial coordinates, depth), which is different from the \code{gulf} package's 'collapse.repeats' function. Experiment codes other than "1" and "5" will be removed when adjusting paired sets. See ?gulf::experiment.str() for details. Default is TRUE.}
}
\value{
A data frame
}
\description{
Uses set cards and length cards. Returns catch numbers and catch weights (kg).
}
\examples{
#library(gulf)
#df<-get_catch(42,2022:2023)
#df<-get_catch(42,1991:2023)
}
\seealso{
gulf::read.surveys(survey = "rv")), gulf::experiment.str()
}
