% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ziff.R
\name{get_ziff}
\alias{get_ziff}
\title{Get raw commercial landings data from a ZIFF (Zonal Interchange Format File)}
\usage{
get_ziff(
  years,
  nafo = NULL,
  species.caught = NULL,
  species.sought = NULL,
  headers = "english"
)
}
\arguments{
\item{years}{Which year(s) to return.}

\item{nafo}{Which NAFO division(s). Case insensitive.}

\item{species.caught}{STATAC species code.}

\item{species.sought}{STATAC species code.}

\item{headers}{Should returned data frame column names (headers) be in English or French. Default is English.}
}
\value{
A data frame
}
\description{
Retrieves data housed at DFO Quebec Region
}
\examples{
#df<-get_ziff(2020,c('4r','4s'),100,100)
}
