% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catch.R
\name{get_catch}
\alias{get_catch}
\title{Get species catch number and catch weight per tow from sGSL September surveys}
\usage{
get_catch(
  species,
  years,
  adj.ratio = TRUE,
  adj.distance = TRUE,
  adj.vessel = TRUE,
  adj.diel = TRUE,
  experiment.code = c(1, 5),
  cutoff.length = NULL
)
}
\arguments{
\item{species}{Which RV species code. Currently limited certain species if applying diel and vessel conversions (10,12,23,40,41,42,43,60,70).}

\item{years}{Which year(s) to return. Current minimum to apply all conversions is 2004.}

\item{adj.ratio}{Adjust for sampling ratio? Default is TRUE.}

\item{adj.distance}{Adjust for target tow distance? Default is TRUE.}

\item{adj.vessel}{Adjust for vessel effects? Default is TRUE.}

\item{adj.diel}{Adjust for diel effects? Default is TRUE. Note that diel adjustments are only made after vessel adjustments to Teleost.}

\item{experiment.code}{Restrict tow experiment type. See gulf::experiment.str(). Default is c(1,5) ("stratified random survey set", "comparative fishing experiment").}

\item{cutoff.length}{Include estimates below and above a cutoff length. Output columns labeled "small" are estimates below the cutoff length. Output columns labeled "big" are estimates at and above the cutoff length. Default is NULL}
}
\value{
A data frame
}
\description{
Uses set cards and length cards.
}
\examples{
#library(gulf)
#df<-get_catch(42,2022:2023)
#df<-get_catch(42,2000:2023)
}
\seealso{
gulf::read.surveys(survey = "rv")), gulf::experiment.str()
}
